Consolidate island logic in `src/islands` folder
