+++ src/islands/shared.tsx
@@ -1,0 +1,5 @@
+import Counter from '../components/Counter.js';
+import TodoApp from '../components/TodoApp.js';
+
+// register islands
+export const registeredIslands = [Counter, TodoApp];
